Refactor to separate users, groups, and factors
