+++ src/okta/constants.ts
@@ -1,0 +1,6 @@
+export const USER_GROUP_ENTITY_TYPE = 'okta_user_group';
+export const APP_USER_GROUP_ENTITY_TYPE = 'okta_app_user_group';
+export const MFA_DEVICE_ENTITY_TYPE = 'mfa_device';
+export const DATA_ACCOUNT_ENTITY = 'DATA_ACCOUNT_ENTITY';
+export const SERVICE_ENTITY_TYPE = 'okta_service';
+export const SERVICE_ENTITY_CLASS = ['Service', 'Control'];
