+++ src/steps/index.ts
@@ -1,7 +1,15 @@
 import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { userSteps } from './users';
+import { groupSteps } from './groups';
+import { factorSteps } from './factors';
 import { applicationSteps } from './applications';
 
-const integrationSteps = [...accountSteps, ...accessSteps, ...applicationSteps];
+const integrationSteps = [
+  ...accountSteps,
+  ...userSteps,
+  ...groupSteps,
+  ...factorSteps,
+  ...applicationSteps,
+];
 
 export { integrationSteps };
